Complete EIP-3155 state test traces

+++ src/Nethermind/Nethermind.Test.Runner/StateTestTxTraceEntry.cs
@@ -27,7 +27,7 @@ namespace Nethermind.Test.Runner
 
         public int Depth { get; set; }
 
-        public int Refund { get; set; }
+        public long Refund { get; set; }
 
         [JsonPropertyName("opName")]
         public string? OperationName { get; set; }
